Создание карточки вынесено в функцию
